# docs/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "NodeHub -  API 文档\n\n这是 NodeHub 的 API 文档\n\n## 认证\n大多数接口需要使用 JWT 令牌进行认证。\n认证时，请在 Authorization 头中包含 JWT 令牌：\n`Authorization: Bearer <your-jwt-token>`\n\n## 错误响应\n所有错误响应都遵循统一格式，包含 code、message 和 error 字段。\n\n## 成功响应\n所有成功响应都遵循统一格式，包含 code、message 和 data 字段。",
    "title": "NodeHub API",
    "contact": {
      "name": "NodeHub API 支持",
      "email": "support@nodehub.com"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://opensource.org/license/gpl-3-0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/setting": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "获取系统所有配置项，支持按分组过滤和关键字搜索",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "配置"
        ],
        "summary": "获取配置项",
        "parameters": [
          {
            "type": "string",
            "description": "分组名称",
            "name": "group",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_setting.Setting"
                      }
                    }
                  }
                }
              ]
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "根据请求数据中的ID批量更新配置项的值和描述",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "配置"
        ],
        "summary": "更新配置项",
        "parameters": [
          {
            "description": "更新配置项请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_setting.Setting"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "github_com_evilCYH_nodehub_internal_models_setting.Setting": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
...
    "github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 200
        },
        "data": {},
        "message": {
          "type": "string",
          "example": "success"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "类型为 \"Bearer\"，后跟空格和 JWT 令牌。",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}